move abstraction out to MCPHandlers
