+++ packages/backend/src/systemTools.ts
@@ -1,4 +1,4 @@
-import { MCPClient } from "@tambo-ai-cloud/core";
+import { MCPClient, MCPHandlers } from "@tambo-ai-cloud/core";
 import OpenAI from "openai";
 
 /** Registry of available MCP tools during a request */
@@ -7,6 +7,8 @@ export interface McpToolRegistry {
   mcpToolsSchema: OpenAI.Chat.Completions.ChatCompletionTool[];
   /** A mapping of MCP tool name to MCP client */
   mcpToolSources: Record<string, MCPClient>;
+
+  mcpHandlers: MCPHandlers;
 }
 
 /** Registry of all browser-side client tools */
